requirements.txt
pyTelegramBotAPI
python-dotenv
rapidfuzz
